Issue #7 and #8 resolved successfully

+++ src/main/resources/db/schema.sql
@@ -17,7 +17,7 @@ SET @OLD_SQL_MODE = @@SQL_MODE, SQL_MODE =
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `products`
 (
-    `id`          INT                     NOT NULL AUTO_INCREMENT,
+    `id`          BIGINT                     NOT NULL AUTO_INCREMENT,
     `name`        VARCHAR(255)            NOT NULL COMMENT 'Nombre del producto',
     `price`       DECIMAL(11, 2) UNSIGNED NOT NULL COMMENT 'Precio del producto',
     `img_url`     VARCHAR(255)            NOT NULL COMMENT 'Url de la imagen del producto.',
@@ -36,7 +36,7 @@ CREATE TABLE IF NOT EXISTS `products`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `users`
 (
-    `id`         INT          NOT NULL AUTO_INCREMENT,
+    `id`         BIGINT          NOT NULL AUTO_INCREMENT,
     `username`   VARCHAR(45)  NOT NULL,
     `password`   VARCHAR(60)  NOT NULL,
     `email`      VARCHAR(100) NULL,
@@ -59,7 +59,7 @@ CREATE TABLE IF NOT EXISTS `shopping_list`
     `total_products` INT UNSIGNED            NOT NULL DEFAULT 0 COMMENT 'Número total de productos.',
     `total_calories` DECIMAL(11, 2) UNSIGNED NOT NULL DEFAULT 0 COMMENT 'Total de calorías de todos los productos.',
     `total_price`    DECIMAL(11, 2) UNSIGNED NOT NULL DEFAULT 0 COMMENT 'Precio total de los productos.',
-    `user_id`        INT                     NOT NULL,
+    `user_id`        BIGINT                     NOT NULL,
     `created_at`     TIMESTAMP               NOT NULL DEFAULT CURRENT_TIMESTAMP,
     `updated_at`     TIMESTAMP               NOT NULL DEFAULT CURRENT_TIMESTAMP,
     PRIMARY KEY (`id`),
@@ -94,7 +94,7 @@ CREATE TABLE IF NOT EXISTS `unit_types`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `products_has_shopping_list`
 (
-    `product_id`        INT                     NOT NULL,
+    `product_id`        BIGINT                     NOT NULL,
     `shopping_list_id`  INT                     NOT NULL,
     `unit_type_id`      INT                     NOT NULL,
     `units_per_product` INT UNSIGNED            NOT NULL DEFAULT 1 COMMENT 'Unidades por producto.',
